scanner middleware: perform minimal changes to links, in bulk
Rather than removing all links, and adding them all again, we extract
the current links, compared them to what we want and only apply the
changes. We do this in bulk with a single call to Django for additions
and removals.

As an additional safety — and possible optimisation — we wrap the
entire call in a transaction.