Removed an odd block of code from WebScanLastModifiedCheckMiddleware
This code tried, if there was no Last-Modified date in the header *and* the
page's content has already been downloaded, to extract a date from a
<meta http-equiv="last-modified" /> tag in the document (a tag which,
incidentally, is not included in any HTML standard). But... why would we
want to check if we should download a document if we've already done so?