[#35209] Restructured WebSource.handles()'s URL enqueueing code
There's now a nested function called handle_url() that should be called
once for each URL encountered; this function will decide whether or not
the URL is new and therefore enqueueable. Moving this logic to a new
function also makes it possible to add support for HTTP 3xx redirects,
which this commit also does!