[#35546] Add env var for manual migrations
Since we need to start two services from the same image, we cannot
always apply migrations through the endpoint; if the two services
are started (almost) simultaneously, the migrations are run twice
which will cause errors.

To preserve the default behaviour, migrations will only be skipped
if a specific environment variable is defined when starting the
container.