[#37386] Remove utils as coverage source
Splitting the test suite has led to an issue with measuring code
coverage for the common utils package, since it is used by all the
main modules, but none of them test it completely.

This led to a more thorough look at the utils package itself, and
the conclusion is to reorganize the content into the three main
modules and removing the utils package as described in [#37444].

Since the package is to be removed in the near future, we have
opted to solve the coverage issue by removing the package as a
coverage source.