[#37367] Move TOML reader to projects folder
For lack of a better placement, the utility functions to read
settings from a toml file has been moved to "projects" as it is
common to both django modules and uses django specific imports.
The latter makes it unsuitable for the Engine, and was introduced
when using toml for the Engine module was dropped (for now).