[#38336] Fix pip-tool Dockerfile
A sys-requirements file for the bash (non-docker) install has
been added (see commit 1dd37bca, related to #38232). This file
is now ignored when building the Dockerfile for running pip-tools,
as the extra requirements do not pertaint to the application
functionality of OS2datascanner itself, but related services such
as postgres and reabbitmq, and the systemd wrapping for the
stages of the engine pipeline.