[#39051] Fix settings export to templates
When the DEBUG setting is incorrectly exported to templates, the
value cannot be accessed from these. A missing value evaluates to
false in a boolean check, such as the one we make to determine
which js bundle to use:
- when DEBUG is true, we use the auto-reloaded regular bundle
- when DEBUG is false, we use the minimized production bundle

Thus, when DEBUG is incorrectly exported to templates, and thus
missing, auto-reloading the frontend in development does not work.