[#37547] Change html converter from BS4 to lxml
BS4 builds a complete tree of the xml. We are only interested in
extracting the body text. Tests shows lxml to be faster(lxml also builds
a tree)
The timing of 5 conversions of a long webpages results in
|   bs4 |     lxml |
|-------+----------|
| 1.008 | 0.060507 |

The converted text is returned as `" ".join(text.split())`, thus all
newlines are stripped. This is done to keep with the old way BS4 was
implemented.
As a note for future, if we want to preserve newlines:
Use `wdiff -3 <file1> <file2> | colordiff` to compare files,
word-by-word.