[#45892] Use regex to match blacklist
We need a new and different approach as we are about the expand our
blacklist and cannot perform that many `str.__contains__`.

Other constraints that make alternative solutions hard are that the new
list will contain strings with spaces and that all the words can often
be followed by ".", ":", or ".:" making `str.split` unviable.