[#46528] Drastically lower the index storage threshold
By default, as an optimisation, messages under 4K get stored in the
RabbitMQ message index and not in the message store. Unfortunately,
virtually all OS2datascanner messages are under 4K, so we were badly
overloading the index, creating performance and reliability problems. This
commit lowers that threshold to 64 bytes, which should mean that all
non-trivial messages get stored properly