[#47229] Allow Sources to declare that their job is to find other Sources
Consider our Microsoft Graph implementation, for example. One Source
traverses the organisational directory and yields Handles for each account
that it finds there; the only thing we can do with these is to build new
Sources based on them. If the system is more aware of that, then it should
be able to route these Handles more efficiently