[#56540] Fix memory leak in user error log search
[#56540] Small fix for the highlightSearch function

The htmx.onLoad-function now scopes the search to only the content being loaded in.

The highlightSearch-function will no longer replace the html-elements it has itself inserted.

All of the page content in error_log.html is now replaced when the form is changed to avoid weird interaction when removing focus from search bar.